+++ preprocessor.h
@@ -6,22 +6,22 @@
 class preprocessor
 {
 public:
-    virtual uint64_t get_preprocessed(uint32_t sample) = 0;
-    virtual uint64_t get_reference() = 0;
+    virtual uint32_t get_preprocessed(uint32_t sample) = 0;
+    virtual uint32_t get_reference() = 0;
     virtual ~preprocessor() = default;
 };
 
 class unit_delay_preprocesson : public preprocessor
 {
 private:
-	int64_t max_value;
-    int64_t min_value = 0;
+	uint32_t max_value;
+    uint32_t min_value = 0;
 	unsigned int sample_size;
-    int64_t preceeding_sample = 0;
-    uint64_t map(int64_t error);
+    uint32_t preceeding_sample = 0;
+    uint32_t map(int64_t error);
 public:
     unit_delay_preprocesson(unsigned int sample_size);
-    uint64_t get_preprocessed(uint32_t sample) override;
-    uint64_t get_reference() override;
+    uint32_t get_preprocessed(uint32_t sample) override;
+    uint32_t get_reference() override;
 	~unit_delay_preprocesson() override = default;
 };
